+++ rootfs/etc/cont-init.d/99-clean.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/with-contenv sh
+
+# Unset sensitive vars
+unset APP_KEY \
+  DB_USERNAME \
+  DB_PASSWORD \
+  REDIS_PASSWORD \
+  PUSHER_APP_SECRET \
+  ANONADDY_SECRET \
+  ANONADDY_SIGNING_KEY_FINGERPRINT